backend/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
configparser==5.3.0
Django==4.2
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.5.0
psycopg==3.1.8
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
